requirements.txt
aiohttp==3.6.3
aiomysql==0.0.20
async-timeout==3.0.1
attrs==20.2.0
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
idna==2.10
idna-ssl==1.1.0
multidict==4.7.6
pkg-resources==0.0.0
pycparser==2.20
PyMySQL==0.9.2
PyYAML==5.3.1
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.5.1
